refactor(docs): extract live finished chart from getting-started guide
Move the 838-point fold, resize, and near-viewport load into
LessonFinishedChart so the guide stays page structure. Reuse
observeNearViewport (same path as theme specimens) and keep the
.finished-chart root for progressive-search VR. Register the new
component as docs content-only so CI does not open VR for prose-only
edits.
